+++ src/app/_layout.tsx
@@ -42,7 +42,12 @@ export default function RootLayout() {
                     />
                     <Stack.Screen
                         name="pages/excerpt/create"
-                        options={{ headerShown: false }}
+                        options={{
+                            headerShown: false,
+                            presentation: "transparentModal",
+                            contentStyle: { backgroundColor: "transparent" },
+                            animation: "none",
+                        }}
                     />
                     <Stack.Screen
                         name="pages/note/[id]"